_backup/Backend_root_clutter/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn==0.24.0.post1

# Database
sqlalchemy==2.0.23
asyncpg==0.31.0
alembic==1.13.1

# HTTP Client (for Telegram API)
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==8.3.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.14.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Pydantic (use compatible version)
pydantic==2.4.2
pydantic-settings==2.0.3

# File Processing (for extraction integration)
pdfminer.six==20231228
pdfplumber==0.9.0
tika==2.6.0

# Optional: For production
gunicorn==21.2.0
